Move PagesTab to its own file
